KTOR-9592 Update the Create a website tutorial and snippets project
