+++ codeSnippets/settings.gradle.kts
@@ -120,7 +120,6 @@ module("snippets", "deployment-ktor-plugin")
 module("snippets", "tutorial-website-static")
 module("snippets", "server-websockets-sharedflow")
 module("snippets", "tutorial-client-get-started")
-module("snippets", "tutorial-server-web-application")
 module("snippets", "server-websockets-serialization")
 module("snippets", "client-websockets-serialization")
 module("snippets", "resource-routing")
