+++ codeSnippets/snippets/tutorial-server-web-application/gradle.properties
@@ -1,0 +1,9 @@
+#Kotlin
+kotlin.code.style=official
+kotlin.daemon.jvmargs=-Xmx8g
+
+#Gradle
+org.gradle.jvmargs=-Xmx12g -Dfile.encoding=UTF-8
+org.gradle.configuration-cache=true
+org.gradle.caching=true
+
