+++ codeSnippets/snippets/tutorial-server-web-application/gradle/libs.versions.toml
@@ -1,0 +1,10 @@
+[versions]
+kotlin = "2.3.21"
+logback = "1.5.21"
+
+[libraries]
+logback-classic = { module = "ch.qos.logback:logback-classic", version.ref = "logback" }
+
+[plugins]
+kotlin-jvm = { id = "org.jetbrains.kotlin.jvm", version.ref = "kotlin" }
+kotlin-multiplatform = { id = "org.jetbrains.kotlin.multiplatform", version.ref = "kotlin" }
