+++ codeSnippets/snippets/tutorial-server-web-application/settings.gradle.kts
@@ -1,0 +1,21 @@
+rootProject.name = "ktor-task-web-app"
+
+pluginManagement {
+    repositories {
+        mavenCentral()
+        gradlePluginPortal()
+    }
+}
+
+dependencyResolutionManagement {
+    repositories {
+        mavenCentral()
+    }
+    versionCatalogs {
+        create("ktorLibs").from("io.ktor:ktor-version-catalog:3.5.0")
+    }
+}
+
+plugins {
+    id("org.gradle.toolchains.foojay-resolver-convention") version "1.0.0"
+}
