--- codeSnippets/snippets/tutorial-server-web-application/src/main/kotlin/Application.kt
@@ -1,15 +1,0 @@
-package com.example
-
-import io.ktor.server.application.*
-import io.ktor.server.engine.*
-import io.ktor.server.netty.*
-
-fun main() {
-    embeddedServer(Netty, port = 8080, host = "0.0.0.0", module = Application::module)
-        .start(wait = true)
-}
-
-fun Application.module() {
-    configureTemplating()
-    configureRouting()
-}
