+++ codeSnippets/snippets/tutorial-server-web-application/src/main/kotlin/Thymeleaf.kt
@@ -1,0 +1,17 @@
+package com.example
+
+import io.ktor.server.application.*
+import io.ktor.server.response.*
+import io.ktor.server.thymeleaf.Thymeleaf
+import io.ktor.server.thymeleaf.ThymeleafContent
+import org.thymeleaf.templateresolver.ClassLoaderTemplateResolver
+
+fun Application.configureThymeleaf() {
+    install(Thymeleaf) {
+        setTemplateResolver(ClassLoaderTemplateResolver().apply {
+            prefix = "templates/thymeleaf/"
+            suffix = ".html"
+            characterEncoding = "utf-8"
+        })
+    }
+}
