+++ codeSnippets/snippets/tutorial-server-web-application/src/main/kotlin/main.kt
@@ -1,0 +1,8 @@
+package com.example
+
+import io.ktor.server.engine.*
+import io.ktor.server.application.*
+
+fun main(args: Array<String>) {
+    io.ktor.server.netty.EngineMain.main(args)
+}
