+++ codeSnippets/snippets/tutorial-server-web-application/src/main/resources/application.yaml
@@ -1,6 +1,7 @@
 ktor:
-    application:
-        modules:
-            - com.example.ApplicationKt.module
-    deployment:
-        port: 8080
+  deployment:
+    port: 8080
+  application:
+    modules:
+      - com.example.ThymeleafKt.configureThymeleaf
+      - com.example.RoutingKt.configureRouting
