+++ codeSnippets/snippets/tutorial-server-web-application/src/main/resources/index.html
@@ -1,0 +1,11 @@
+<!DOCTYPE html >
+<html xmlns:th="http://www.thymeleaf.org">
+<head>
+<meta charset="UTF-8">
+<title>Title</title>
+</head>
+<body>
+<span th:text="${user.name}"></span>
+</body>
+</html>
+
