--- codeSnippets/snippets/tutorial-server-web-application/src/test/kotlin/ApplicationTest.kt
@@ -1,21 +1,0 @@
-package com.example
-
-import io.ktor.client.request.get
-import io.ktor.client.statement.bodyAsText
-import io.ktor.http.HttpStatusCode
-import io.ktor.server.testing.testApplication
-import kotlin.test.Test
-import kotlin.test.assertEquals
-
-class ApplicationTest {
-    @Test
-    fun testRoot() = testApplication {
-        application {
-            configureRouting()
-        }
-        client.get("/").apply {
-            assertEquals(HttpStatusCode.Companion.OK, status)
-            assertEquals("Hello World!", bodyAsText())
-        }
-    }
-}
