+++ codeSnippets/snippets/tutorial-server-web-application/src/test/kotlin/ServerTest.kt
@@ -1,0 +1,16 @@
+import io.ktor.client.request.get
+import io.ktor.http.HttpStatusCode
+import io.ktor.server.testing.testApplication
+import kotlin.test.*
+
+class ServerTest {
+
+    @Test
+    fun `test root endpoint`() = testApplication {
+        // loads default configuration
+        configure()
+        // verify server root returns 200
+        assertEquals(HttpStatusCode.OK, client.get("/").status)
+    }
+
+}
